rename run stream
