+++ src/extensions/flows/run/run.cmd.tsx
@@ -6,8 +6,8 @@ import { ReplaySubject } from 'rxjs';
 import { Command, CLIArgs } from '../../cli';
 import { Flags, PaperOptions } from '../../paper';
 import { Flows } from '../flows';
-import { reportRunStream } from './handle-run-stream';
-import { flowEvents } from './handle-run-stream';
+import { reportRunStream } from './report-run-stream';
+import { flowEvents } from './report-run-stream';
 import { Report } from './report';
 import { Reporter } from '../../reporter';
 import { Logger } from '../../logger';
